Aleks liss changed test

+++ src/test/java/ru/aleks/weather/repository/JdbcSessionRepositoryTest.java
@@ -2,11 +2,13 @@ package ru.aleks.weather.repository;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ActiveProfiles;
 import ru.aleks.weather.config.H2DatabaseConfig;
+import ru.aleks.weather.exception.SessionNotFoundException;
 import ru.aleks.weather.mapper.SessionMapper;
 import ru.aleks.weather.mapper.UserMapper;
 import ru.aleks.weather.model.Session;
@@ -26,6 +28,7 @@ class JdbcSessionRepositoryTest {
     private static JdbcSessionRepository jdbcSessionRepository;
     private static JdbcUserRepository jdbcUserRepository;
     private static final UUID uuid = UUID.randomUUID();
+    private User user;
 
     @BeforeAll
     public static void setUp() {
@@ -33,63 +36,57 @@ class JdbcSessionRepositoryTest {
         DataSource dataSource = h2DatabaseConfig.dataSource();
         JdbcTemplate jdbcTemplate = h2DatabaseConfig.jdbcTemplate(dataSource);
         jdbcSessionRepository = new JdbcSessionRepository(
-            jdbcTemplate, new SessionMapper()
+                jdbcTemplate, new SessionMapper()
         );
         jdbcUserRepository = new JdbcUserRepository(
                 jdbcTemplate, new UserMapper()
         );
+    }
 
+    @BeforeEach
+    public void save() {
+        if (jdbcUserRepository.getByLogin("vova").isEmpty()) {
+            user = jdbcUserRepository.save(new User("vova", "123")).get();
+        }
     }
 
     @AfterEach
     public void clear() {
-        jdbcUserRepository.deleteByLogin("vova");
-        jdbcSessionRepository.delete(uuid);
+        if (user.getLogin() != null) {
+            jdbcUserRepository.deleteByLogin("vova");
+        }
     }
 
     @Test
     public void whenSavedSessionThenReturnSavedSession() {
-        Optional<User> savedUser = jdbcUserRepository.getByLogin("vova");
-        if (savedUser.isEmpty()) {
-            User user = new User("vova", "123");
-            savedUser = jdbcUserRepository.save(user);
-        }
-        assertThat(savedUser).isPresent();
         Session session = new Session(
                 uuid,
-                savedUser.get().getId(),
+                user.getId(),
                 LocalDateTime.now()
         );
         Optional<Session> savedSession = jdbcSessionRepository.save(session);
         assertThat(savedSession).isPresent();
-        assertThat(savedSession.get().getUserId()).isEqualTo(savedUser.get().getId());
     }
 
     @Test
     public void whenSavedSessionAndGetItByIdThenReturnSavedSession() {
-        Optional<User> savedUser = jdbcUserRepository.getByLogin("vova");
-        if (savedUser.isEmpty()) {
-            User user = new User("vova", "123");
-            savedUser = jdbcUserRepository.save(user);
-        }
-        assertThat(savedUser).isPresent();
         Session session = new Session(
                 uuid,
-                savedUser.get().getId(),
+                user.getId(),
                 LocalDateTime.now()
         );
         Optional<Session> savedSession = jdbcSessionRepository.save(session);
         assertThat(savedSession).isPresent();
         Optional<Session> foundSession = jdbcSessionRepository.getById(uuid);
         assertThat(foundSession).isPresent();
-        assertThat(foundSession.get().getUserId()).isEqualTo(savedUser.get().getId());
+        assertThat(foundSession.get().getUserId()).isEqualTo(user.getId());
     }
 
 
     @Test
     public void whenDontSavedSessionAndGetItByIdThenThrowException() {
         assertThatThrownBy(() -> jdbcSessionRepository.getById(UUID.randomUUID()))
-                .isInstanceOf(RuntimeException.class)
+                .isInstanceOf(SessionNotFoundException.class)
                 .hasMessage("session was not found");
     }
 }
